Fix for training continuation if an argument is a tuple.
When JSON serializes a tuple and then deserializes it, it becomes a
list. This list then is not considered equal to the tuple defined in
argumnts.py. This happend for the newly added argument
conv_embed_num_filters which basically broke training continuation, as
Sockeye thought one of the arguments differed.