dump in utf-8
The current code dumps to ASCII, encoding all UTF-8 characters, like so:

    [snip]
    "\u2581,": 6,
    "\u2581.": 7,
    "\u2581de": 8,
    [snip]

This makes the file a bit hard to read. This change dumps the file directly to UTF-8. I have tested that loading the dictionary dumped this way produces an identical dictionary to one loaded from the ASCII version.