Training resumption with non-empty model dir. (#146)
Do not refuse training when no model state is found. So far the training
just died if the model directory already existed but did not contain
a training state. I do not see any good reason to not just start training
from scratch from the existing directory.

Additionally this commit adds a nicer error message for decoding when no
parameter file is found in the model directory.