Added hard lexical constraints (Post & Vilar, NAACL 2018)
This commit adds hard lexical constraints on top of current master.
It also introduces a number of changes to beam search that were important
to getting lexical constraints to work, namely:

- Only applying normalization to completed hypotheses
- Adding pruning (see PR #230)

These changes are redundant with PR #230, and should be merged together
at some point.

To use the constraints, the decoder requires JSON input in the following form:

    { "text": "Einer soll ein hoch@@ rangi@@ ges Mitglied aus Berlin gewesen sein .",
      "constraints": ["is said to", "powerful"] }

BPE must have been applied to both the text and the constraints. A convenience script
is included to compute this:

    echo -e "text\tconstraint 1\tconstraint2..." | python3 -m sockeye.lexical_constraints --bpe BPE_MODEL

This can be passed in directly to sockeye.translate if the `--json-input` flag is used.