allow `--tokenize none`
This may be useful with a custom test set which is already tokenized.
Either the detokenized version of the reference may not be available
or you want to explore what is the effect of de/tokenization
e.g. when replicating an experiment from a paper
which used multi-bleu.perl on tokenized output
(which is a bad practice, I know).