added negative constraints (#433)
This implements negative constraints in beam search, i.e., phrases that are not permitted to appear in the output. It works by tracking the progress of each negative constraint; if the last word is reached, the corresponding entry in the scores array is set to np.inf prior to the call to topk(). Both a global negative constraint list (via `--avoid-list /path/to/file`) and a sentence-specific list (via the JSON object) are supported.