removed unnecessary "fill_up" policy
This was added to mark the end of an incomplete batch. But this can be accomplished much more easily by setting the `pad` element of `mxnet.io.DataBatch` (which exists for exactly this purpose). Thus we can get rid of the fill_up policy entirely, which is nice, because it was added just for scoring. Also, this is necessary in order to handle printing an output for invalid sentences in the middle of a batch.