rewrote scoring to load data on demand and not skip sentences (#593)
The current scoring implementation (a) loads the entire training data into memory and (b) skips sentences that are too long so that the output is not parallel with the input. This PR fixes both of these issues. Data is loaded from disk one batch at a time via a new `BatchRawParallelSampleIter`, and sentences that are too long are truncated.

One application of this is [dual cross-entropy filtering](http://aclweb.org/anthology/W18-6478) of the training data.