Quantization: use tied weight parameter
MXNet 1.6 introduced optional deduplication when files are saved.  This effectively relies on the same weight objects being shared when a model is loaded, but previously quantization broke that.

This breaks existing models quantized to int8 format but is consistent with its own quantize.py.  There aren't many of those models out there though.